Update bt-sink-switch.py
To add an option in bt-sink-switch.py that allows you to set the force_switch parameter to True or False, you can use the argparse library. Here is a modified version of your script
Now you can call the script with the --force-switch argument to set the parameter to True.